+++ tmsBackEnd/src/main/java/com/yash/tms/services/MasterVehicleDetailsManager.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.yash.tms.entity.MasterVehicleDetails;
 
 public interface MasterVehicleDetailsManager {
-	public List<MasterVehicleDetails> findAllVehicleDetails(short vehicleDetailsIsDeleted);
+	public List<MasterVehicleDetails> findAllVehicleDetails(int vehicleDetailsIsDeleted);
 
 	public MasterVehicleDetails addVehicleDetails(MasterVehicleDetails vehicleDetails);
